final refactored functions
