+++ backend-services/configuration/core/configuration.py
@@ -3,7 +3,6 @@ import multiprocessing as mp
 import os
 import re
 import shutil
-import socket
 import stat
 import time
 from io import StringIO
@@ -36,7 +35,6 @@ from artemis_utils.constants import MITIGATION_HOST
 from artemis_utils.constants import NOTIFIER_HOST
 from artemis_utils.constants import PREFIXTREE_HOST
 from artemis_utils.constants import RIPERISTAP_HOST
-from artemis_utils.envvars import COMPOSE_PROJECT_NAME
 from artemis_utils.envvars import IS_KUBERNETES
 from artemis_utils.envvars import RABBITMQ_URI
 from artemis_utils.envvars import REDIS_HOST
@@ -46,6 +44,9 @@ from artemis_utils.rabbitmq import create_exchange
 from artemis_utils.rabbitmq import create_queue
 from artemis_utils.redis import ping_redis
 from artemis_utils.redis import redis_key
+from artemis_utils.service import get_local_ip
+from artemis_utils.service import service_to_ips_and_replicas_in_compose
+from artemis_utils.service import service_to_ips_and_replicas_in_k8s
 from artemis_utils.translations import translate_as_set
 from artemis_utils.translations import translate_asn_range
 from artemis_utils.translations import translate_rfc2622
@@ -91,60 +92,6 @@ MONITOR_SERVICES = [
     BGPSTREAMHISTTAP_HOST,
     EXABGPTAP_HOST,
 ]
-
-
-# need to move to utils
-def get_local_ip():
-    return socket.gethostbyname(socket.gethostname())
-
-
-# need to move to utils
-def service_to_ips_and_replicas(base_service_name):
-    local_ip = get_local_ip()
-    service_to_ips_and_replicas_set = set([])
-    addr_infos = socket.getaddrinfo(base_service_name, REST_PORT)
-    for addr_info in addr_infos:
-        af, sock_type, proto, canon_name, sa = addr_info
-        replica_ip = sa[0]
-        # do not include yourself
-        if base_service_name == SERVICE_NAME and replica_ip == local_ip:
-            continue
-        replica_host_by_addr = socket.gethostbyaddr(replica_ip)[0]
-        replica_name_match = re.match(
-            r"^"
-            + re.escape(COMPOSE_PROJECT_NAME)
-            + r"_"
-            + re.escape(base_service_name)
-            + r"_(\d+)",
-            replica_host_by_addr,
-        )
-        replica_name = "{}-{}".format(base_service_name, replica_name_match.group(1))
-        service_to_ips_and_replicas_set.add((replica_name, replica_ip))
-    return service_to_ips_and_replicas_set
-
-
-# need to move to utils
-def service_to_ips_and_replicas_in_k8s(base_service_name):
-    from kubernetes import client, config
-
-    service_to_ips_and_replicas_set = set([])
-    config.load_incluster_config()
-    current_namespace = open(
-        "/var/run/secrets/kubernetes.io/serviceaccount/namespace"
-    ).read()
-    v1 = client.CoreV1Api()
-    try:
-        endpoints = v1.read_namespaced_endpoints_with_http_info(
-            base_service_name, current_namespace, _return_http_data_only=True
-        ).to_dict()
-        for entry in endpoints["subsets"][0]["addresses"]:
-            replica_name = entry["target_ref"]["name"]
-            replica_ip = entry["ip"]
-            service_to_ips_and_replicas_set.add((replica_name, replica_ip))
-    except Exception as e:
-        log.exception(e)
-
-    return service_to_ips_and_replicas_set
 
 
 def read_conf(load_yaml=True, config_file=None):
@@ -792,7 +739,9 @@ def post_configuration_to_other_services(
             if IS_KUBERNETES:
                 ips_and_replicas = service_to_ips_and_replicas_in_k8s(service)
             else:
-                ips_and_replicas = service_to_ips_and_replicas(service)
+                ips_and_replicas = service_to_ips_and_replicas_in_compose(
+                    SERVICE_NAME, service
+                )
         except Exception:
             log.error("could not resolve service '{}'".format(service))
             continue
