+++ backend-services/mitigation/core/mitigation.py
@@ -4,16 +4,15 @@ import time
 from typing import Dict
 from typing import NoReturn
 
-import requests
 import ujson as json
 from artemis_utils import get_logger
 from artemis_utils.constants import DATABASE_HOST
-from artemis_utils.constants import HEALTH_CHECK_TIMEOUT
 from artemis_utils.constants import PREFIXTREE_HOST
 from artemis_utils.envvars import RABBITMQ_URI
 from artemis_utils.envvars import REST_PORT
 from artemis_utils.rabbitmq import create_exchange
 from artemis_utils.rabbitmq import create_queue
+from artemis_utils.service import wait_data_worker_dependencies
 from kombu import Connection
 from kombu import Producer
 from kombu import uuid
@@ -31,38 +30,6 @@ shared_memory_locks = {"data_worker": mp.Lock()}
 # global vars
 SERVICE_NAME = "mitigation"
 DATA_WORKER_DEPENDENCIES = [PREFIXTREE_HOST, DATABASE_HOST]
-
-
-# need to move this to utils
-def wait_data_worker_dependencies(data_worker_dependencies):
-    while True:
-        met_deps = set()
-        unmet_deps = set()
-        for service in data_worker_dependencies:
-            try:
-                r = requests.get(
-                    "http://{}:{}/health".format(service, REST_PORT),
-                    timeout=HEALTH_CHECK_TIMEOUT,
-                )
-                status = True if r.json()["status"] == "running" else False
-                if not status:
-                    unmet_deps.add(service)
-                else:
-                    met_deps.add(service)
-            except Exception:
-                unmet_deps.add(service)
-        if len(unmet_deps) == 0:
-            log.info(
-                "all needed data workers started: {}".format(data_worker_dependencies)
-            )
-            break
-        else:
-            log.info(
-                "'{}' data workers started, waiting for: '{}'".format(
-                    met_deps, unmet_deps
-                )
-            )
-        time.sleep(1)
 
 
 class ConfigHandler(RequestHandler):
