+++ backend-services/prefixtree/core/prefixtree.py
@@ -53,7 +53,6 @@ shared_memory_locks = {
 SERVICE_NAME = "prefixtree"
 
 
-# need to move this to artemis-utils
 def pytricia_to_dict(pyt_tree):
     pyt_dict = {}
     for prefix in pyt_tree:
@@ -61,7 +60,6 @@ def pytricia_to_dict(pyt_tree):
     return pyt_dict
 
 
-# need to move this to artemis-utils
 def dict_to_pytricia(dict_tree, size=32):
     pyt_tree = pytricia.PyTricia(size)
     for prefix in dict_tree:
